+++ base/constants/commons.go
@@ -1,6 +1,7 @@
 package constants
 
 const (
-	DateFormat      = "02-01-2006 15:04:05"
-	DevEnvironments = "dev,staging"
+	DateFormat         = "02-01-2006 15:04:05"
+	DateFormatDateTime = "2006-01-02T15:04:05"
+	DevEnvironments    = "dev,staging"
 )